requirements.txt
selenium>=4.0.0
openai>=1.0.0
gspread>=5.7.0
oauth2client>=4.1.3
python-dotenv>=1.0.0
pillow>=9.0.0
